# docs/api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "vaultaffe",
    "description": "The HTTP surface of one vaultaffe instance, and the only way in or out of it (Specification §9). Every instance is at its own address. Everything but the handshake and the problem catalogue lives under /api/v1/, and every refusal is an application/problem+json document whose `code` names it.",
    "version": "v1"
  },
  "paths": {
    "/api/handshake": {
      "get": {
        "tags": [
          "Contract"
        ],
        "summary": "What this instance is and which versions of the contract it serves.",
        "operationId": "Handshake",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Handshake"
                }
              }
            }
          }
        }
      }
    },
    "/problems": {
      "get": {
        "tags": [
          "Contract"
        ],
        "summary": "Every refusal this instance can make, by code.",
        "operationId": "ReadProblems",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProblemKind"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/problems/{code}": {
      "get": {
        "tags": [
          "Contract"
        ],
        "summary": "One refusal, by code.",
        "operationId": "ReadProblem",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemKind"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/instance": {
      "get": {
        "tags": [
          "Identity"
        ],
        "summary": "Whether this instance has been started, and what its organization is called.",
        "operationId": "ReadInstance",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Instance"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Identity"
        ],
        "summary": "The first run: the default organization, and the first user as its administrator.",
        "operationId": "StartInstance",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartInstanceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstanceStarted"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions": {
      "post": {
        "tags": [
          "Identity"
        ],
        "summary": "Email and password in, a session token out. The token is shown exactly once.",
        "operationId": "SignIn",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignInRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions/current": {
      "delete": {
        "tags": [
          "Identity"
        ],
        "summary": "Revoke the session this request came in under.",
        "operationId": "SignOut",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/me": {
      "get": {
        "tags": [
          "Identity"
        ],
        "summary": "The caller: the organization, the person, and the token it came in under.",
        "operationId": "ReadMe",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Me"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/device/authorizations": {
      "post": {
        "tags": [
          "Device login"
        ],
        "summary": "Begin a login: a short code for the human, a long one for the client.",
        "operationId": "BeginDeviceLogin",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceLogin"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/device/tokens": {
      "post": {
        "tags": [
          "Device login"
        ],
        "summary": "Poll: the session token once a human has confirmed, a code that says why not until then.",
        "operationId": "RedeemDeviceLogin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RedeemDeviceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tokens": {
      "post": {
        "tags": [
          "Tokens"
        ],
        "summary": "Create a service or agent token. Its value appears here and nowhere else, ever.",
        "operationId": "CreateToken",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenIssued"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Tokens"
        ],
        "summary": "Every token of this organization, newest first, revoked ones included.",
        "operationId": "ReadTokens",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Token"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tokens/{id}": {
      "delete": {
        "tags": [
          "Tokens"
        ],
        "summary": "Revoke a token. Revoked rather than deleted, so its entries keep an author.",
        "operationId": "RevokeToken",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Binding": {
        "required": [
          "projectId",
          "environmentId"
        ],
        "type": "object",
        "properties": {
          "projectId": {
            "type": "string",
            "format": "uuid"
          },
          "environmentId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          }
        }
      },
      "CreateTokenRequest": {
        "required": [
          "kind",
          "name"
        ],
        "type": "object",
        "properties": {
          "kind": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "scopes": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "bindings": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/Binding"
            }
          },
          "expiresAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
      "DeviceLogin": {
        "required": [
          "deviceCode",
          "userCode",
          "verificationUri",
          "verificationUriComplete",
          "expiresInSeconds",
          "intervalSeconds"
        ],
        "type": "object",
        "properties": {
          "deviceCode": {
            "type": "string"
          },
          "userCode": {
            "type": "string"
          },
          "verificationUri": {
            "type": "string"
          },
          "verificationUriComplete": {
            "type": "string"
          },
          "expiresInSeconds": {
            "type": "integer",
            "format": "int32"
          },
          "intervalSeconds": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "Handshake": {
        "required": [
          "product",
          "release",
          "apiVersions",
          "minimumClient"
        ],
        "type": "object",
        "properties": {
          "product": {
            "type": "string"
          },
          "release": {
            "type": "string"
          },
          "apiVersions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "minimumClient": {
            "type": "string"
          }
        }
      },
      "Instance": {
        "required": [
          "started",
          "organizationName"
        ],
        "type": "object",
        "properties": {
          "started": {
            "type": "boolean"
          },
          "organizationName": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "InstanceStarted": {
        "required": [
          "organizationId",
          "organizationName",
          "session"
        ],
        "type": "object",
        "properties": {
          "organizationId": {
            "type": "string",
            "format": "uuid"
          },
          "organizationName": {
            "type": "string"
          },
          "session": {
            "$ref": "#/components/schemas/Session"
          }
        }
      },
      "Me": {
        "required": [
          "organizationId",
          "userId",
          "name",
          "isAdministrator",
          "tokenId",
          "tokenName",
          "tokenKind",
          "scopes"
        ],
        "type": "object",
        "properties": {
          "organizationId": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "isAdministrator": {
            "type": "boolean"
          },
          "tokenId": {
            "type": "string",
            "format": "uuid"
          },
          "tokenName": {
            "type": [
              "null",
              "string"
            ]
          },
          "tokenKind": {
            "type": "string"
          },
          "scopes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "ProblemDetails": {
        "required": [
          "code"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "code": {
            "type": "string"
          }
        }
      },
      "ProblemKind": {
        "required": [
          "code",
          "status",
          "title"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          }
        }
      },
      "RedeemDeviceRequest": {
        "required": [
          "deviceCode"
        ],
        "type": "object",
        "properties": {
          "deviceCode": {
            "type": "string"
          }
        }
      },
      "Session": {
        "required": [
          "userId",
          "name",
          "email",
          "isAdministrator",
          "token",
          "expiresAt"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "isAdministrator": {
            "type": "boolean"
          },
          "token": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "SignInRequest": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "StartInstanceRequest": {
        "required": [
          "email",
          "name",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "Token": {
        "required": [
          "id",
          "kind",
          "name",
          "scopes",
          "bindings",
          "reachesTheWholeOrganization",
          "createdAt",
          "expiresAt",
          "revokedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "type": "string"
          },
          "name": {
            "type": [
              "null",
              "string"
            ]
          },
          "scopes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "bindings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Binding"
            }
          },
          "reachesTheWholeOrganization": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "expiresAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "revokedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
      "TokenIssued": {
        "required": [
          "token",
          "value"
        ],
        "type": "object",
        "properties": {
          "token": {
            "$ref": "#/components/schemas/Token"
          },
          "value": {
            "type": "string"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Contract"
    },
    {
      "name": "Identity"
    },
    {
      "name": "Device login"
    },
    {
      "name": "Tokens"
    }
  ]
}